+++ Palace/Book/Models/TPPBook.swift
@@ -68,7 +68,11 @@ let UpdatedKey: String = "updated"
   static func categoryStringsFromCategories(categories: [TPPOPDSCategory]) -> [String] {
     categories.filter { $0.scheme == nil || ($0.scheme?.absoluteString ?? "") == SimplifiedScheme }.map { $0.label ?? $0.term }
   }
-  
+
+  @objc var isAudiobook: Bool {
+    defaultBookContentType == .audiobook
+  }
+
   init(
     acquisitions: [TPPOPDSAcquisition],
     authors: [TPPBookAuthor]?,
@@ -117,7 +121,7 @@ let UpdatedKey: String = "updated"
     self.contributors = contributors
     self.bookTokenLock = NSRecursiveLock()
   }
-
+  
   /// @brief Factory method to build a TPPBook object from an OPDS feed entry.
   ///
   /// @param entry An OPDS entry to base the book on.
@@ -224,8 +228,7 @@ let UpdatedKey: String = "updated"
         relatedBooksURL: authorLinkStrings.count > index ? URL(string: authorLinkStrings[index]) : nil
       )
     }
-    
-    
+  
     var revokeURL = URL(string: dictionary[RevokeURLKey] as? String ?? "")
     var reportURL = URL(string: dictionary[ReportURLKey] as? String ?? "")
