+++ Palace/Book/UI/TPPBookDownloadingCell.m
@@ -165,7 +165,6 @@
 - (void)didSelectCancel
 {
   [self.delegate didSelectCancelForBookDownloadingCell:self];
-  [self.delegate didSelectCancelForBookDownloadingCell:self];
 }
 
 @end
